Fix NPE and tangible package name
